[python] Allow to pass optional pythonName to algorithm.load()
The used downstream generator uses an argument
`pythonName` which seems to be only used in
logging output. So far `dune.generator.algorithm.load()`
always passes the function signature as `pythonName`.

When jit-compiling functions with many arguments
and complicated type names, this may provide very
verbose output. On the other hand, it's very well
possible that in many instances of generated
functions the signature coincides, while the
significant difference is in the actual function
body rending the verbose output useless.

This patch adds support for passing a custom value
for `pythonName` to `dune.generator.algorithm.load()`
by a corresponding keyword argument. If the argument
is not passed, the signature is used as before.