Fix private config header file generate info with current module
This is mainly an oversight on 45fa74f02a09d77002c7ceedeb7e7b275d432135. The dune module information was being updated before the public part but not before the private part. This caused private header part to be generated with other module information (e.g. dune-grid would be populated with dune-common variables)