Implement TransformedRangeIterator using IteratorFacade
This comes with a slight change: The old helper
`Dune::Impl::PointerProxy` was dropped in favor of the new
`Dune::ProxyArrowResult`. The latter no longer supports storing
raw references. This change is deliberately: In the no-proxy
case, we can simply use raw pointers instead of the helper
to avoid indirection and keep simple cases simple.

Furthermore it this make `TransformedRangeIterator` simpler
and allows to make it more flexible:

* All iterator operations despite dereferencing
  are now autogenerated, by exporting `baseIterator()`.
* Allow to either store a function or a pointer to function.
  The transformation is evaluated conditionally, if a call
  `f(z)` is valid, this is used, else `(*f)(z)` is used.
* Use `Dune::AutonomousValue` instead of `std::decay_t`
  to deduce `value_type`. This improves support for types
  using this mechanism.
* Store `reference` instead of `value_type` in `ProxyArrowResult`.
  This is important if both types are different and the `reference`
  proxy e.g. forwards to some other storage.
* Add checks for free `TransformedRangeIterator` storing functions
  with value and pointer semantics.