[concept] Simplify tupleEntriesModel()
This helper allows to check if all entries from a tuple model
a concept. The patch simplifes the check and reduces
the dependencies: We no longer depend on `tupleutility.hh`
which is pretty heavy weight since in includes `hybridutilities.hh`
and `rangeutilities.hh`. Thus emulated concepts could not be used
in these headers so far, which is fixed by the patch.

To prevent a regression this also adds a test for the previously
untestet function.