[bugfix] Fix typedef for dereferencing an ConstArrayListIterator
The type returned when dereferencing an iterator should
correspond to the typedef named `reference` even if the iterator
acts on a const container. For the `ArrayList::const_iterator`
this was not the case so far and `reference` was a non-const `value_type&`
which is fixed by this patch. While not beeing part of the standard
iterator interface, the old typedef `const_reference` is kept for
backward compatibility and consistency with `ArrayList::iterator`.