[dunecontrol] Skip directories without CMake config files for module_DIR
find_package(module) will fail if the directory set with module_DIR
does not contain CMake configuration file.

Hence the current approach requires that there are pkgconfig files
installed and found for each module to be used.

With this change we fall back to installed modules if the directory
does not contain any Cmake configuration files.

This fixes problems with using OPM as DUNE modules (in Debian) which
stopped shipping broken pkgconf file 2023-09