[dunecontrol] Module without CMakeLists.txt next to dune.module is installed
The decision whether a module is installed is usually based on whether
pkgconfig was able to find it. If the DUNE module does not ship  a
pkgconf file or pkgconf is missing, then dunecontrol tries to build
this installed modules, too. That fails miserably.

With this commit we add a check at the end that checks whether
there is a CMakeLists.txt file in the source directory of the
module. This directory is where the dune.module file was found. If
there is none we also assume that the module is installed and do not
try to build it.