[ci] Make .gitlab-ci.yml portable
Currently the CI config requires that the `core/ci-config`
project exists on the same gitlab server. This prevents
the CI from working when pushing the repository to another
server. While one may argue, that one also has to push
`core/ci-config` then, it is a very strong restriction to
require that the gitlab group `core` is available.

This can be avoided by using a gitlab remote include for those
files with a link pointing to the dune server instead of a local
include.