Implement helper types IteratorFacadeForTraits and DefaultIteratorTraits
While the `Dune::IteratorFacade` mixin-class was deliberately
designed to have in interface that mimics the old iterator facade
classes, passing all respective types as separate template parameters,
it is sometimes more convenient to pass them using a single traits
class - especially since the standard already provides us with the
design of this traits class, namely `std::iterator_traits`.

This patch provides this interface by introducing a convenience alias

```cpp
template<class IteratorImpl, class IteratorTraits>
using IteratorFacadeForTraits = Dune::IteratorFacade<
      IteratorImpl,
      typename IteratorTraits::iterator_category,
      typename IteratorTraits::value_type,
      typename IteratorTraits::reference,
      typename IteratorTraits::pointer,
      typename IteratorTraits::difference_type>;
```

Additionally it provides a default implementation of an `iterator_traits`
class. This is possible, since you can in most cases derive all respective
types from a single one. To support proxy classes this is not
`value_type` but `reference`. The default traits is

```cpp
template<class IteratorCategory, class Reference, class DifferenceType = std::ptrdiff_t>
struct DefaultIteratorTraits
{
  using iterator_category = IteratorCategory;
  using reference = Reference;
  using value_type = Dune::AutonomousValue<reference>;
  using pointer = std::conditional_t<std::is_lvalue_reference_v<reference>, value_type*, Dune::ProxyArrowResult<reference>>;
  using difference_type = DifferenceType;
};
```

Notice that the main intention of this class is not to safe typing but
to aid the user in providing the correct types even in case of proxy
iterators. The latter is non-trivial as can be seen from the fact that
we didn't get it correct with the old iterator facades ourselves.