[bugfix] Fix a post in/de-crement of IteratorFacade
The `IteratorFacade` mixin class allows to implement the increment
in the derived class in several ways:

1. By manually providing `operator++()`
2. By providing access to a base-iterator that should be incremented
3. By manually providing `operator+=(...)`

(same for decrementing). While versions 2 and 3 work as expected,
version 1 does not, if only the prefix `operator++()` but not the
postfix `operator++(int)` is implemented in the derived class, because
the former shaddows `operator++(int)` in the base class. As a remedy
we can declare `using IteratorFacade::operator++` in the derived
class to provide access to the base class version. This however,
does not work because it currently does not rely on the overwritten
`operator++()` in the derived class.

This patch fixes both issues: It mentions the need for the `using`
statement in the documentation and let's the base class
`operator++(int)` call `operator++()` via the derived class.

The same applies for `operator--(int)` analogouesly.