[doc] Fix changelog entry on DP_DEFAULT_INCLUDE_DIRS
The changelog addresses users and developers. Hence we should
not tell them that they have to `provide a utility ...`
but when they have to use it. Furthermore the default value
of the policy in 2.11 and (even more important) in 2.12
was missing - despite the fact that the latter is a breaking
change. The fact that `dune_target_enable_all_packages`
now means 'all except for the present one' was also
added.